refactor: restart vad engine from a clean state + separate more tests + more tets of sub componentas
