+++ screenpipe-audio/Cargo.toml
@@ -70,6 +70,9 @@ realfft = "3.4.0"
 regex = "1.11.0"
 
 
+rodio = { version = "0.19.0", optional = true, features = ["wav"] }
+
+
 [target.'cfg(target_os = "windows")'.dependencies]
 ort = { version = "2.0.0-rc.5", features = [
   "download-binaries",
@@ -80,6 +83,8 @@ ort = { version = "2.0.0-rc.5", features = [
 esaxx-rs = "0.1.10"
 samplerate = { version = "0.2.4" }
 libsamplerate-sys = "0.1.10"
+
+
 
 [target.'cfg(target_os = "macos")'.dependencies]
 once_cell = "1.17.1"
@@ -95,7 +100,7 @@ futures = "0.3.31"
 tracing-subscriber = "0.3.16"
 tokio-tungstenite = "0.24.0"
 futures-util = "0.3.31"
-rodio = "0.19.0"
+# rodio = "0.19.0"
 rand = "0.8.5"
 indicatif = "0.17.8"
 
@@ -103,6 +108,7 @@ indicatif = "0.17.8"
 metal = ["candle/metal", "candle-nn/metal", "candle-transformers/metal"]
 cuda = ["candle/cuda", "candle-nn/cuda", "candle-transformers/cuda"]
 mkl = ["candle/mkl", "candle-nn/mkl", "candle-transformers/mkl"]
+test-utils = ["rodio"]
 
 
 [[bench]]
