+++ screenpipe-audio/src/constants.rs
@@ -36,21 +36,10 @@ pub struct Config {
     pub overlap_samples: usize,
 
     // VAD settings
-    pub frame_history: usize,
-    pub speech_threshold: f32,
-    pub silence_threshold: f32,
-    pub speech_frame_threshold: usize,
-    pub chunk_size: usize,
-    pub speech_duration_threshold_ms: u64,
-    pub silence_duration_threshold_ms: u64,
-
-    // VAD sensitivity settings
-    pub vad_sensitivity_low_threshold: f32,
-    pub vad_sensitivity_medium_threshold: f32,
-    pub vad_sensitivity_high_threshold: f32,
-    pub vad_sensitivity_low_speech_ratio: f32,
-    pub vad_sensitivity_medium_speech_ratio: f32,
-    pub vad_sensitivity_high_speech_ratio: f32,
+    pub pre_speech_buffer_duration_secs: f32,
+    pub vad_buffer_duration_secs: f32,
+    pub speech_threshold_duration_ms: u64,
+    pub silence_threshold_duration_ms: u64,
 
     // Add a name field to identify configurations
     pub name: String,
@@ -101,119 +90,46 @@ impl Config {
             overlap_samples: overlap_seconds * whisper_sample_rate as usize,
 
             // VAD settings
-            frame_history: env::var("VAD_FRAME_HISTORY")
+            pre_speech_buffer_duration_secs: env::var("VAD_PRE_SPEECH_BUFFER_DURATION_SECS")
                 .ok()
                 .and_then(|v| v.parse().ok())
-                .unwrap_or(15),
-            speech_threshold: env::var("VAD_SPEECH_THRESHOLD")
+                .unwrap_or(2.0),
+            vad_buffer_duration_secs: env::var("VAD_BUFFER_DURATION_SECS")
                 .ok()
                 .and_then(|v| v.parse().ok())
-                .unwrap_or(0.55),
-            silence_threshold: env::var("VAD_SILENCE_THRESHOLD")
-                .ok()
-                .and_then(|v| v.parse().ok())
-                .unwrap_or(0.3),
-            speech_frame_threshold: env::var("VAD_SPEECH_FRAME_THRESHOLD")
-                .ok()
-                .and_then(|v| v.parse().ok())
-                .unwrap_or(4),
-            chunk_size: env::var("VAD_CHUNK_SIZE")
-                .ok()
-                .and_then(|v| v.parse().ok())
-                .unwrap_or(1600),
-            speech_duration_threshold_ms: env::var("VAD_SPEECH_DURATION_THRESHOLD_MS")
+                .unwrap_or(1.0),
+            speech_threshold_duration_ms: env::var("VAD_SPEECH_THRESHOLD_DURATION_MS")
                 .ok()
                 .and_then(|v| v.parse().ok())
                 .unwrap_or(700),
-            silence_duration_threshold_ms: env::var("VAD_SILENCE_DURATION_THRESHOLD_MS")
+            silence_threshold_duration_ms: env::var("VAD_SILENCE_THRESHOLD_DURATION_MS")
                 .ok()
                 .and_then(|v| v.parse().ok())
                 .unwrap_or(1500),
-
-            // VAD sensitivity settings
-            vad_sensitivity_low_threshold: env::var("VAD_SENSITIVITY_LOW_THRESHOLD")
-                .ok()
-                .and_then(|v| v.parse().ok())
-                .unwrap_or(0.25),
-            vad_sensitivity_medium_threshold: env::var("VAD_SENSITIVITY_MEDIUM_THRESHOLD")
-                .ok()
-                .and_then(|v| v.parse().ok())
-                .unwrap_or(0.55),
-            vad_sensitivity_high_threshold: env::var("VAD_SENSITIVITY_HIGH_THRESHOLD")
-                .ok()
-                .and_then(|v| v.parse().ok())
-                .unwrap_or(0.75),
-            vad_sensitivity_low_speech_ratio: env::var("VAD_SENSITIVITY_LOW_SPEECH_RATIO")
-                .ok()
-                .and_then(|v| v.parse().ok())
-                .unwrap_or(0.02),
-            vad_sensitivity_medium_speech_ratio: env::var("VAD_SENSITIVITY_MEDIUM_SPEECH_RATIO")
-                .ok()
-                .and_then(|v| v.parse().ok())
-                .unwrap_or(0.07),
-            vad_sensitivity_high_speech_ratio: env::var("VAD_SENSITIVITY_HIGH_SPEECH_RATIO")
-                .ok()
-                .and_then(|v| v.parse().ok())
-                .unwrap_or(0.3),
 
             name,
         }
     }
 
-    // Add factory methods for different configurations
-    #[cfg(test)]
-    pub fn generate_test_configs() -> Vec<Self> {
-        vec![
-            Self {
-                name: "quick_detection".to_string(),
-                speech_threshold: 0.6,
-                silence_threshold: 0.3,
-                speech_duration_threshold_ms: 500, // Detect speech faster (500ms)
-                silence_duration_threshold_ms: 1000, // End segments faster (1s)
-                overlap_seconds: 2,                // 2s overlap for context
-                chunk_duration_ms: 2000.0,         // 2s chunks
-                ..Self::new("quick_detection".to_string())
-            },
-            Self {
-                name: "careful_detection".to_string(),
-                speech_threshold: 0.5,
-                silence_threshold: 0.35,
-                speech_duration_threshold_ms: 1000, // More careful speech detection (1s)
-                silence_duration_threshold_ms: 2000, // Longer silence needed (2s)
-                overlap_seconds: 3,                 // 3s overlap for safety
-                chunk_duration_ms: 3000.0,          // 3s chunks
-                ..Self::new("careful_detection".to_string())
-            },
-            Self {
-                name: "max_context".to_string(),
-                speech_threshold: 0.45,
-                silence_threshold: 0.4,
-                speech_duration_threshold_ms: 700, // Default speech detection
-                silence_duration_threshold_ms: 1500, // Default silence
-                overlap_seconds: 4,                // Maximum context preservation
-                chunk_duration_ms: 4000.0,         // Longer chunks
-                ..Self::new("max_context".to_string())
-            },
-            Self {
-                name: "minimal_overlap".to_string(),
-                speech_threshold: 0.55,
-                silence_threshold: 0.3,
-                speech_duration_threshold_ms: 600,
-                silence_duration_threshold_ms: 1200,
-                overlap_seconds: 1,        // Minimum overlap
-                chunk_duration_ms: 2500.0, // Standard chunk size
-                ..Self::new("minimal_overlap".to_string())
-            },
-        ]
-    }
-
-    #[cfg(test)]
+    #[cfg(any(test, feature = "test-utils"))]
     pub fn set_as_active(self) -> Arc<Self> {
         let config = Arc::new(self);
         TESTING_CONFIG.with(|test_config| {
             *test_config.borrow_mut() = Some(config.clone());
         });
         config
+    }
+
+    #[cfg(any(test, feature = "test-utils"))]
+    pub fn speech_threshold_duration_ms(mut self, duration: u64) -> Self {
+        self.speech_threshold_duration_ms = duration;
+        self
+    }
+
+    #[cfg(any(test, feature = "test-utils"))]
+    pub fn silence_threshold_duration_ms(mut self, duration: u64) -> Self {
+        self.silence_threshold_duration_ms = duration;
+        self
     }
 }
